New benchmarks

+++ Salar.BinaryBuffers/BinarySpanBufferWriter.cs
@@ -390,6 +390,7 @@ public ref struct BinarySpanBufferWriter: IBufferWriter
 		var length = buffer.Length;
 		Advance(length);
 
+		// Slice(pos) was 31-145% slower and produced larger code in .NET 10 benchmarks.
 		buffer.CopyTo(_buffer.Slice(pos, length));
 	}
 
@@ -433,7 +434,7 @@ public ref struct BinarySpanBufferWriter: IBufferWriter
 	/// Moves the current position of the writer ahead by the specified number of bytes.
 	/// </summary>
 	/// <param name="count">The number of bytes to advance</param>
-	/// <exception cref="EndOfStreamException"/>
+	/// <exception cref="System.IO.EndOfStreamException"/>
 	[MethodImpl(MethodImplOptions.AggressiveInlining)]
 	public void SimulateWrite(int count)
 	{
@@ -454,7 +455,7 @@ public ref struct BinarySpanBufferWriter: IBufferWriter
 		_position = newPos;
 
 		// Fast branchless/simplified tracker update:
-		if (newPos > _writtenLength)
+		if ((uint)newPos > (uint)_writtenLength)
 		{
 			_writtenLength = newPos;
 		}
